+++ src/blas/sparse/crs_spmv.cpp
@@ -14,7 +14,7 @@ namespace monolish{
 		logger.func_in(monolish_func);
 
 		//err
-		if( x.size() != y.size() || A.get_row() != (int)x.size()){
+		if( x.size() != y.size() || A.get_row() != (size_t)x.size()){
 			throw std::runtime_error("error vector size is not same");
 
 		}
@@ -26,21 +26,21 @@ namespace monolish{
 		double* yd = y.data();
 
 		double* vald = A.val.data();
-		int* rowd = A.row_ptr.data();
-		int* cold = A.col_ind.data();
+		size_t* rowd = A.row_ptr.data();
+		size_t* cold = A.col_ind.data();
 
 		#pragma acc data pcopyin(xd[0:n], vald[0:nnz], rowd[0:n+1], cold[0:nnz]) copyout(yd[0:n]) 
 		{
 			#pragma acc kernels
 			{
 				#pragma acc loop independent 
-				for(int i = 0 ; i < n; i++){
+				for(size_t i = 0 ; i < n; i++){
 					yd[i] = 0;
 				}
 
 				#pragma acc loop independent
-				for(int i = 0 ; i < n; i++){
-					for(int j = rowd[i] ; j < rowd[i+1]; j++){
+				for(size_t i = 0 ; i < n; i++){
+					for(size_t j = rowd[i] ; j < rowd[i+1]; j++){
 						yd[i] += vald[j] * xd[cold[j]];
 					}
 				}
@@ -50,12 +50,12 @@ namespace monolish{
 #else // cpu
 
 	#pragma omp parallel for 
-		for(int i = 0 ; i < A.get_row(); i++)
+		for(size_t i = 0 ; i < A.get_row(); i++)
 			y.val[i] = 0;
 
 	#pragma omp parallel for
-		for(int i = 0 ; i < A.get_row(); i++)
-			for(int j = A.row_ptr[i] ; j < A.row_ptr[i+1]; j++)
+		for(size_t i = 0 ; i < A.get_row(); i++)
+			for(size_t j = A.row_ptr[i] ; j < A.row_ptr[i+1]; j++)
 				y.val[i] += A.val[j] * x.val[A.col_ind[j]];
 
 #endif
